requirements.txt
Cython
numpy
#pkg-resources
PyOpenGL
PyQt5
PyQt5-sip
scipy
triangle
